requirements.txt
# Core AI/ML Libraries
dspy-ai>=2.4.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
accelerate>=0.24.0

# CLIP for Multimodal Embeddings (included in transformers)
# CLIP models loaded via transformers library
ftfy>=6.1.0  # Text normalization for CLIP
regex>=2023.0.0  # Required by CLIP

# RAG and Vector Database
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
faiss-cpu>=1.7.4

# Document Processing
PyMuPDF>=1.23.0  # PDF processing with text and images
python-docx>=1.1.0  # DOCX files
openpyxl>=3.1.0  # XLSX files
python-pptx>=0.6.21  # PPTX support

# Image Processing
Pillow>=10.0.0
pytesseract>=0.3.10  # OCR for images
opencv-python>=4.8.0

# Web UI
streamlit>=1.29.0
streamlit-chat>=0.1.1

# Utilities
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.5.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
chardet>=5.2.0
